refactor: use inline testing with vitest
