+++ tsconfig.json
@@ -17,7 +17,10 @@
     "paths": {
       "@/*": ["./src/*"],
       "~/*": ["./playground/*"]
-    }
+    },
+    "types": [
+      "vitest/importMeta"
+    ]
   },
   "include": ["src", "playground", "tests"]
 }
